requirements.txt
together>=0.1.5
python-dotenv>=0.19.0
pandas>=1.3.0
sqlalchemy>=1.4.0
numpy>=1.20.0
flask>=2.0.0
requests>=2.25.0
google-cloud-bigquery>=3.0.0
scikit-learn>=1.0.0
